--- a/datashader/glyphs/line.py
+++ b/datashader/glyphs/line.py
@@ -9,7 +9,7 @@ from datashader.utils import isnull, isreal, ngjit
 from numba import cuda
 import numba.types as nb_types
 
-from .._dependencies import cudf, cupy as cp, spd
+from .._dependencies import cudf, cupy as cp
 
 from ..transfer_functions._cuda_utils import cuda_args