--- a/datashader/tiles.py
+++ b/datashader/tiles.py
@@ -8,7 +8,7 @@ import os
 
 import numpy as np
 
-from ._dependencies import dask, db
+from ._dependencies import dask
 
 __all__ = ['render_tiles', 'MercatorTileDefinition']
 
@@ -51,6 +51,7 @@ def calculate_zoom_level_stats(super_tiles, load_data_func,
         if is_bool:
             span = (0, 1)
         elif dask:
+            import dask.bag as db
             b = db.from_sequence(stats)
             span = dask.compute(b.min(), b.max())
         else:
@@ -65,6 +66,9 @@ def render_tiles(full_extent, levels, load_data_func,
                  post_render_func, output_path, color_ranging_strategy='fullscan'):
     if not dask:
         raise ImportError('Dask is required for rendering tiles')
+
+    import dask.bag as db
+
     results = {}
     for level in levels:
         print('calculating statistics for level {}'.format(level))