… datashader/datashape/tests/test_discovery.py
… datashader/transfer_functions/_cuda_utils.py
… datashader/utils.py
… pyproject.toml