Trying to see if I could lazy import modules, so we don't get such a heavy import cost of datashader. 

This is heavily inspired by the work done in https://github.com/holoviz/holoviews/pull/6476, also here an import hook for when the module is imported. Though, I have chosen to have the import here in a centralized file as this matches better with how datashader is structured. 

Nowhere near finishing in the transition, also currently dask stalls and don't know why...

--- 
# Benchmark
## Hyperfine

### Main
``` bash
❯ hyperfine "python -c ''" "python -c 'import datashader'" --warmup 5
Benchmark 1: python -c ''
  Time (mean ± σ):      17.1 ms ±   1.0 ms    [User: 12.9 ms, System: 4.0 ms]
  Range (min … max):    14.8 ms …  21.0 ms    172 runs

Benchmark 2: python -c 'import datashader'
  Time (mean ± σ):      1.276 s ±  0.026 s    [User: 2.710 s, System: 0.182 s]
  Range (min … max):    1.249 s …  1.339 s    10 runs

Summary
  python -c '' ran
   74.79 ± 4.69 times faster than python -c 'import datashader'
```

### This branch
``` bash
❯ hyperfine "python -c ''" "python -c 'import datashader'" --warmup 5
Benchmark 1: python -c ''
  Time (mean ± σ):      17.2 ms ±   1.0 ms    [User: 12.6 ms, System: 4.4 ms]
  Range (min … max):    14.7 ms …  19.5 ms    164 runs

Benchmark 2: python -c 'import datashader'
  Time (mean ± σ):     820.2 ms ±  11.7 ms    [User: 2305.5 ms, System: 123.2 ms]
  Range (min … max):   803.0 ms … 848.5 ms    10 runs
```

## Tuna
```
❯ python -X importtime -c 'import datashader' 2> tuna.log && tuna tuna.log
```
Ignore the `__version` time as this does not affect users. 

### Main
![image](https://github.com/user-attachments/assets/05313cda-ae3a-4788-bc9a-d81e63a9b58c)

### This branch

![image](https://github.com/user-attachments/assets/d6e23614-9a31-42ba-a1b2-44b395312ff7)