@jbednar 2025-05-14
Cool, thanks. Note that there are at least two sources of long startup time for Datashader, i.e. imports and Numba compilation. I think only once Numba can be precompiled and/or cached (which is _supposed_ to be possible already) will the startup time be reasonable.