@hoxbro 2025-05-15
> Cool, thanks. Note that there are at least two sources of long startup time for Datashader, i.e. imports and Numba compilation. I think only once Numba can be precompiled and/or cached (which is supposed to be possible already) will the startup time be reasonable.

Yes, it could be great also to tackle that, but just by looking at the profiler, I can see, e.g., `dask.dataframe` accounted for half of the import time, and pandas a little below  2/10.  

If I lazy load pandas (and therefore xarray and comment out some ragged extension inheritance for now), I can get it down to the following: 

``` python
❯ hyperfine "python -c ''" "python -c 'import datashader'" --warmup 5
Benchmark 1: python -c ''
  Time (mean ± σ):       9.6 ms ±   0.7 ms    [User: 6.6 ms, System: 2.9 ms]
  Range (min … max):     7.7 ms …  11.8 ms    277 runs

Benchmark 2: python -c 'import datashader'
  Time (mean ± σ):     298.7 ms ±  14.9 ms    [User: 1847.7 ms, System: 57.0 ms]
  Range (min … max):   284.5 ms … 331.5 ms    10 runs

Summary
  python -c '' ran
   31.19 ± 2.67 times faster than python -c 'import datashader'
```

![image](https://github.com/user-attachments/assets/c5523037-6e53-48ac-a06e-4ea5724f47af)